+++ src/Products/PloneMeeting/migrations/migrate_to_4211.py
@@ -20,11 +20,25 @@ class Migrate_To_4211(Migrator):
                 cfg.setHideHistoryTo(new_values)
         logger.info('Done.')
 
+    def _updateSearchCopyGroupsSearchesCondition(self):
+        """Use MeetingConfig.show_copy_groups_search."""
+        logger.info('Updating copy groups searches TAL condition in every MeetingConfigs...')
+        for cfg in self.tool.objectValues('MeetingConfig'):
+            searchallitemsincopy = cfg.searches.searches_items.get('searchallitemsincopy')
+            if searchallitemsincopy:
+                searchallitemsincopy.tal_condition = 'python: cfg.show_copy_groups_search()'
+            searchunreaditemsincopy = cfg.searches.searches_items.get('searchunreaditemsincopy')
+            if searchunreaditemsincopy:
+                searchunreaditemsincopy.tal_condition = \
+                    'python: cfg.getEnableLabels() and cfg.show_copy_groups_search()'
+        logger.info('Done.')
+
     def run(self, extra_omitted=[], from_migration_to_4200=False):
 
         logger.info('Migrating to PloneMeeting 4211...')
 
         self._migrateConfigHideHistoryTo()
+        self._updateSearchCopyGroupsSearchesCondition()
 
         logger.info('Migrating to PloneMeeting 4211... Done.')
 
@@ -32,7 +46,8 @@ class Migrate_To_4211(Migrator):
 def migrate(context):
     '''This migration function will:
 
-       1) Migrate attribute MeetingConfig.hideHistoryTo for MeetingConfig.
+       1) Migrate attribute MeetingConfig.hideHistoryTo for MeetingConfig;
+       2) Update searchallcopygroups/searchunreaditemsincopy searches tal_condition.
     '''
     migrator = Migrate_To_4211(context)
     migrator.run()